+++ Sachssoft.Sasopuls.Views.Shared/Messaging/IInvalidateable.cs
@@ -1,0 +1,7 @@
+namespace Sachssoft.Sasopuls.Messaging
+{
+    public interface IInvalidateable
+    {
+        void Invalidate();
+    }
+}