+++ Sachssoft.Sasopuls.Views.Shared/Messaging/IInvalidateableCommandRule.cs
@@ -1,0 +1,9 @@
+using System;
+
+namespace Sachssoft.Sasopuls.Messaging
+{
+    public interface IInvalidateableCommandRule
+    {
+        void Invalidate();
+    }
+}